requirements.txt
aniso8601==9.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JSON==0.3.4
Flask-RESTful==0.3.9
flask-restful-swagger-2==0.35
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
Shapely==1.8.5.post1
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0